requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
anthropic>=0.17.0
langchain==0.3.27
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
python-dotenv==1.0.0
pydantic>=2.7.4
